Move helper functions to separate module
